7038-user-edit-history
show true-users in edit history
fix n+1
dry up version behavior into mixin
