+++ app/views/admin/edit_histories/_list.haml
@@ -1,0 +1,35 @@
+:ruby
+  # fetch all the user records in one query
+  user_ids = list.flat_map do |version|
+    [ version.clean_user_id, version.clean_true_user_id]
+  end
+  users_by_id = User.with_deleted.where(id: user_ids.uniq).index_by(&:id)
+
+  system_user = User.system_user
+  version_usernames = {}
+  list.each do |version|
+    if version.whodunnit == 'unauthenticated'
+      version_usernames[version.id] = Hmis::User.system_user
+      next
+    end
+
+    user = users_by_id[version.clean_true_user_id] || users_by_id[version.clean_user_id]
+    version_usernames[version.id] = user ? user&.name.presence : "User ID #{user.id}"
+  end
+
+.table-responsive
+  %table.table.table-striped
+    %thead
+      %tr
+        %th Date
+        %th Editor
+        %th Changes
+    %tbody
+      - list.each do |version|
+        %tr
+          %td= version.created_at.to_fs
+          %td= version_usernames[version.id] || version.whodunnit.presence || 'Anonymous'
+          %td
+            - changes = describe_changes_to(version)&.each do |change|
+              = change
+              %br
